Have articles for each user individually

+++ mine_backend/app/articlesapi/routes.py
@@ -1,10 +1,11 @@
 from bson.errors import InvalidId
 from flask import jsonify, request, abort, Response
 from flask_pymongo import ObjectId
+from flask_jwt_extended import jwt_required, current_user
 
 from app import app, mongo
 from app.articlesapi import bp
-from app.articlesapi.scrapers import TheFlowScraper
+from app.articlesapi.scrapers import TheFlowScraper, UnknownChildType, UnknownTagName
 from app.request_params import get_int_arg
 
 
@@ -23,12 +24,18 @@ def summary():
 
 
 @bp.route('/articles/', methods=['POST'])
+@jwt_required
 def save_article():
     link = request.json.get('link')
     if link:
-        result = mongo.db.saved_articles.insert_one(TheFlowScraper
-                                                    .scrap_article(link)
-                                                    .to_dict())
+        try:
+            article_dict = TheFlowScraper.scrap_article(link).to_dict()
+        except (UnknownTagName, UnknownChildType) as ex:
+            return abort(500, str(ex))
+
+        article_dict['user_id'] = current_user.object_id
+
+        result = mongo.db.saved_articles.insert_one(article_dict)
         return jsonify({
             'insertedId': str(result.inserted_id),
         }), 201
@@ -37,14 +44,16 @@ def save_article():
 
 
 @bp.route('/articles/', methods=['GET'])
+@jwt_required
 def get_articles():
     limit = get_int_arg(request, 'limit', 20)
     offset = get_int_arg(request, 'offset', 0)
 
-    articles = list(mongo.db.saved_articles.find(skip=offset, limit=limit))
+    articles = list(mongo.db.saved_articles.find({'user_id': current_user.object_id}, skip=offset, limit=limit))
 
     for article in articles:
         article['_id'] = str(article['_id'])
+        article['user_id'] = str(article['user_id'])
 
     return jsonify({
         'articles': articles,
@@ -52,17 +61,20 @@ def get_articles():
 
 
 @bp.route('/articles/<id>/', methods=['GET'])
+@jwt_required
 def get_article(id):
-    article = mongo.db.saved_articles.find_one_or_404(ObjectId(id))
+    article = mongo.db.saved_articles.find_one_or_404({'_id': ObjectId(id), 'user_id': current_user.object_id})
     article['_id'] = str(article['_id'])
+    article['user_id'] = str(article['user_id'])
 
     return jsonify(article)
 
 
 @bp.route('/articles/<id>/', methods=['DELETE'])
+@jwt_required
 def delete_article(id):
     try:
-        mongo.db.saved_articles.delete_one({'_id': ObjectId(id)})
+        mongo.db.saved_articles.delete_one({'_id': ObjectId(id), 'user_id': current_user.object_id})
         return Response(status=204)
     except InvalidId as ex:
         return jsonify({
